Handel user and AI messages being added to screen. Stubbed Api Call
